// apps/web/types/api.generated.ts
export interface components {
    schemas: {
// ...
        HttpsystemStatus: {
            embedder: {
                [key: string]: unknown;
            };
            embedding_coverage: {
                [key: string]: number;
            };
            env: string;
            health: {
                [key: string]: unknown;
            };
            jobs: {
                [key: string]: number;
            };
            offline_mode: boolean;
            providers: components["schemas"]["ServiceProviderStatus"][];
            version: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}